+++ tests/resources/test_mgmt.py
@@ -33,7 +33,7 @@ async def test_health(app_client):
     assert body["pgstac"]["pgstac_version"]
 
 
-async def test_health_503(database):
+async def test_health_503(pgstac):
     """Test health endpoint error."""
 
     # No lifespan so no `get_connection` is application state
@@ -58,11 +58,11 @@ async def test_health_503(database):
 
     # No lifespan so no `get_connection` is application state
     postgres_settings = PostgresSettings(
-        pguser=database.user,
-        pgpassword=database.password,
-        pghost=database.host,
-        pgport=database.port,
-        pgdatabase=database.dbname,
+        pguser=pgstac.user,
+        pgpassword=pgstac.password,
+        pghost=pgstac.host,
+        pgport=pgstac.port,
+        pgdatabase=pgstac.dbname,
     )
     # Create connection pool but close it just after
     await connect_to_db(api.app, postgres_settings=postgres_settings)