+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
+- remove pygstac dependency
+- refactor tests fixtures to test multiple version of PgSTAC
+
 ## [6.1.1] - 2025-11-20
 
 ### Fixed